complete payment

+++ app/src/main/java/com/example/testcontentprovider/model/ChiTietGioHang.java
@@ -44,7 +44,7 @@ public class ChiTietGioHang implements Serializable{
         this.soLuong = soLuong;
     }
 
-    public double getDonGia() {
+    public int getDonGia() {
         return donGia;
     }
